📦 chore: max credit 수정
